+++ data-raw/data_example.R
@@ -1,16 +1,16 @@
 # Get Example Data ------------------------------------------------------------
 
 # Create example databases
-bcch <- nc_data_dl(request_id = 152543, username = "sample")
+bcch <- nc_data_dl(request_id = 152543, fields_set = "minimum", username = "sample")
 usethis::use_data(bcch, internal = FALSE, overwrite = TRUE)
 
-hofi <- nc_data_dl(species = 20350, username = "sample", info = "pkg_data")
+hofi <- nc_data_dl(species = 20350, fields_set = "minimum", username = "sample", info = "pkg_data")
 usethis::use_data(hofi, internal = FALSE, overwrite = TRUE)
 
 unlink(file.path("inst", "extdata", "bcch.nc"))
-nc_data_dl(request_id = 152543, username = "sample", 
+nc_data_dl(request_id = 152543, fields_set = "minimum", username = "sample", 
            sql_db = file.path("inst", "extdata", "bcch"))
 
 unlink(file.path("inst", "extdata", "hofi.nc"))
-nc_data_dl(species = 20350, username = "sample", info = "pkg_data",
+nc_data_dl(species = 20350, fields_set = "minimum", username = "sample", info = "pkg_data",
            sql_db = file.path("inst", "extdata", "hofi"))