+++ NEWS.md
@@ -8,6 +8,7 @@
 * Clarify metadata versions as dates
 * Clarify which metadata is stored locally
 * Add checks and warn users if local metadata >4 weeks old
+* Make 'extended' the default field set (#31)
 
 # naturecounts 0.4.1
 * Fix use of species_id and record_id in `cosewic_ranges()`.